modal v2

+++ HealthDiary/HealthDiary_Front/HealthDiary_Front/src/app/components/sidebar/sidebar/sidebar.component.ts
@@ -1,6 +1,7 @@
 import { Component, TemplateRef } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
-import { ModalService } from 'src/app/services/modal-service/modal-service.service';
+import { PopupModalComponent } from 'src/app/helpers/popup-modal/popup-modal/popup-modal.component';
 
 @Component({
   selector: 'app-sidebar',
@@ -10,12 +11,23 @@ import { ModalService } from 'src/app/services/modal-service/modal-service.servi
 export class SidebarComponent {
 
   constructor(private router: Router,
-    private modalService: ModalService){}
+    private dialog: MatDialog) {}
 
-  public logout(modalTemplate: TemplateRef<any>){
-    this.modalService.open(modalTemplate, {body: 'Are you sure you want to logout?', title: 'Logout'}).subscribe(action => {
-       localStorage.clear();
-       this.router.navigate(['/login']);      
-    });
-  }
+    public logout(): void {
+      const dialogRef = this.dialog.open(PopupModalComponent, {
+        data: {
+          modalTitle: 'Logout',
+          modalBody: 'Are you sure you want to logout?'
+        }
+      });
+  
+      dialogRef.afterClosed().subscribe(() => {});
+
+      dialogRef.componentInstance.confirmationEvent.subscribe((result: boolean) => {
+        if (result) {
+          localStorage.clear();
+          this.router.navigate(['/login']);        
+        }
+      });
+    }
 }
